lakebridge/switch/requirements.txt
anytree==2.13.0
chardet==5.2.0
databricks-sdk~=0.67.0
httpx==0.28.1
omegaconf==2.3.0
PyYAML==6.0.2
tenacity==9.1.2
tiktoken==0.9.0
